improve empty checks for common interfaces implementer
