+++ util/gconv/gconv_slice_any.go
@@ -11,11 +11,6 @@ import (
 	"reflect"
 )
 
-// apiInterfaces is used for type assert api for Interfaces.
-type apiInterfaces interface {
-	Interfaces() []interface{}
-}
-
 // SliceAny is alias of Interfaces.
 func SliceAny(i interface{}) []interface{} {
 	return Interfaces(i)
